#715: Adding recommendation Endpoint
